test(harness): enforce one bounded node-startup budget
Starting a test node had no advertised worst case: eight launches, each waiting thirty seconds for readiness and then up to five minutes for the product's shutdown watchdog, bounded one node by forty-four minutes before a cluster's remaining nodes were even started. Every failure retried, including configuration errors, application errors and panics, and the retry restarted the readiness timeout each time.

One absolute budget now covers a node's whole startup. It starts before the first launch, and every readiness probe, every cleanup after a failed attempt and every pause between attempts spends only what is left of it. The budget is derived from what one attempt costs, so raising any part of it moves the advertised worst case rather than hiding inside it, and it is asserted against the ceiling the harness advertises. A cluster is built one node at a time, so its worst case is that budget repeated for every node, and the construction deadline it opens bounds every node it starts.

Startup outcomes are classified before a retry. A readiness deadline, a listener that could not bind and an interconnect transport that could not start are transient and may be launched again on fresh ports while budget remains. A launch the harness could not perform, any other application error, a panic, a cancellation and a clean exit end the startup at once. Cleanup after a failed attempt asks for a graceful stop within a short slice of what remains and aborts and joins the task when that slice ends, so the product's multi-minute shutdown watchdog is never reached once per retry.

Every attempt transition is reported with its typed cause, its classification, its cleanup outcome and the budget left, and exhaustion carries the whole history, so resource pressure and a deterministic failure are told apart from the diagnostic alone.

The constants come from a full scenario-suite run at the CI concurrency factor of two scenarios per CPU: 3,465 node startups, a 2.0s median, a 4.3s 99th percentile and a 24.2s slowest healthy startup, which one attempt outlasts.

+++ tests/common/node_liveness.rs
@@ -285,7 +285,7 @@ impl fmt::Display for LastReadinessOutcome {
 )]
 pub(crate) struct NodeStartupError {
     pub(crate) node: ClusterNodeName,
-    pub(crate) attempt: usize,
+    pub(crate) attempt: u32,
     pub(crate) elapsed: Duration,
     pub(crate) failure: NodeStartupFailure,
     pub(crate) task_state: NodeTaskState,
@@ -295,7 +295,7 @@ pub(crate) struct NodeStartupError {
 impl NodeStartupError {
     fn report(
         node: &ClusterNodeName,
-        attempt: usize,
+        attempt: u32,
         deadline: PhaseDeadline,
         failure: NodeStartupFailure,
         task_state: NodeTaskState,
@@ -320,7 +320,7 @@ impl OwnedNodeTask {
     pub(crate) async fn wait_until_ready<P, Probe>(
         &mut self,
         node: &ClusterNodeName,
-        attempt: usize,
+        attempt: u32,
         deadline: PhaseDeadline,
         poll_interval: Duration,
         mut probe: P,
